firware update added

+++ pi/pyScript.py
@@ -53,13 +53,22 @@ def initiate():
 
     input_state = GPIO.input(18)
     if not input_state:
-        reset()
+        pass# reset()
 
-    control = open('current/control.txt', 'r')
-    c1 = control.readline()[0]
-    c2 = control.readline()[0]
-    dataFile = control.readline()
-    control.close()
+    try:
+        control = open('current/control.txt', 'r')
+        c1 = control.readline()[0]
+        c2 = control.readline()[0]
+        dataFile = control.readline()
+        control.close()
+    except Exception as exp:
+        control = open('current/control.txt', 'r')
+        control.write('$')
+        control.write('\n')
+        control.write('s')
+        c1='$'
+        c2='s'
+        control.close()
 
     # Start the appropriate process on current file
     if c2 == 'v':
@@ -73,8 +82,8 @@ def initiate():
 
 GPIO.setmode(GPIO.BCM)
 GPIO.setup(18, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-GPIO.setup(23,GPIO.OUT)
-GPIO.output(23,1)
+# GPIO.setup(23,GPIO.OUT)
+# GPIO.output(23,1)
 initiate()
 while True:
     try:
@@ -82,13 +91,13 @@ while True:
 
         input_state = GPIO.input(18)
         if not input_state:
-            reset()
+            pass #reset()
 
         if os.path.isfile('buffer/control.txt') :
 
             #Blink 10 times
-            for i in range(0,10):
-                blink()
+            # for i in range(0,10):
+            #     blink()
 
             # Read control file
             control = open('buffer/control.txt', 'r')
@@ -110,8 +119,9 @@ while True:
                 # Delete all current files
                 files = os.listdir('current')
                 for file in files:
-                    if not file == '':
-                        os.system('rm -rf "current/%s"' %(file))
+                    if not file == 'control.txt':
+                        if not file == '':
+                            os.system('rm -rf "current/%s"' %(file))
 
                 # Move data,control files to current
                 if not dataFile == '':
@@ -126,14 +136,14 @@ while True:
                     pass
                 if c2=='i':
                     pqivx = Popen(['pqiv','-i','--fullscreen','current/%s'%(dataFile)])
+                if c2=='u':
+                    Popen('python updateScript.py',shell=True)
+                    # Popen(['python','updateScript.py'])
 
                 # Delete all files in buffer
                 deleteAllBuffer()
 
     except Exception as exp:
-        GPIO.output(18, 0)
+        # GPIO.output(23, 0)
 
         time.sleep(1)
-
-
-
